requirements.txt
pylint==4.0.5
pytest==9.0.2
pytest-cov==7.1.0
black==26.3.1
google-genai==1.68.0